+++ src/config.js
@@ -1,0 +1,4 @@
+/**
+ * Created by sonvu on 15/03/2017.
+ */
+export const ROOT_URL = `http://localhost:4000/api`;
