+++ src/reducers/index.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux';
-import PatientReducer from './reducer_patient';
-import PatientSelected from './reducer_patient_selected';
-
+// import PatientReducer from './reducer_patient';
+// import PatientSelected from './reducer_patient_selected';
+import {fetchPatientReducer, selectPatientReducer, updatePatientReducer} from './patient_reducer';
 const rootReducer = combineReducers({
-  patients: PatientReducer,
-  patientSelected: PatientSelected
+  patients: fetchPatientReducer,
+  patientSelected: selectPatientReducer
 });
 
 export default rootReducer;
